script: Move single usage files into subdirectories
All of these files were only used in a single place.
Therefore move them into those relevant directories
and reduce visibility.

Testing: it compiles
