+++ components/script/dom/document/image_animation.rs
