+++ components/script/dom/window/layout_image.rs
