+++ components/script/event_loop/mod.rs
@@ -8,3 +8,4 @@ mod script_mutation_observers;
 #[expect(unsafe_code)]
 pub(crate) mod script_thread;
 pub(crate) mod script_window_proxies;
+mod svg_font;
