+++ components/script/lib.rs
@@ -34,11 +34,8 @@ mod event_loop;
 pub(crate) mod fetch;
 pub(crate) mod indexeddb;
 mod init;
-mod layout_image;
 mod url;
 
-pub(crate) mod iframe_collection;
-pub(crate) mod image_animation;
 pub mod layout_dom;
 pub(crate) mod messaging;
 mod microtask;
@@ -53,7 +50,6 @@ pub(crate) mod script_runtime;
 pub(crate) mod serviceworker_manager;
 mod stylesheet_loader;
 mod stylesheet_set;
-mod svg_font;
 pub mod test;
 mod timers;
 mod webdriver_handlers;
